add a function comment,including commentCount and its contents

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
 
     /*修改问题*/
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
+    public String edit(@PathVariable(name="id") Long id,
                        Model model){
         QuestionDTO question = questionService.getQuesById(id);
         model.addAttribute("title",question.getTitle());
@@ -45,7 +45,7 @@ public class PublishController {
     public String dopublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) long id,
                             HttpServletRequest request,
                             Model model){
         model.addAttribute("title",title);
