+++ src/main/java/life/majiang/community/controller/QuestionController.java
@@ -17,7 +17,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{quesid}")
-    public String question(@PathVariable(name="quesid") Integer quesid,
+    public String question(@PathVariable(name="quesid") Long quesid,
                            Model model){
         QuestionDTO questionDTO = questionService.getQuesById(quesid);
         //获取成功之后累加阅读数
