+++ src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -3,16 +3,17 @@ package life.majiang.community.dto;
 import life.majiang.community.model.Question;
 import life.majiang.community.model.User;
 import lombok.Data;
+import sun.util.resources.ga.LocaleNames_ga;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer LikeCount;
